+++ src/main/scala/basis/algebra/Ring.scala
@@ -10,15 +10,15 @@ package basis.algebra
 /** An element of a commutative ring. A ring is an abelian group under addition
   * and a semigroup under multiplication. Multiplication distributes over addition.
   * 
-  * $IntMorphismInfo
+  * $LongMorphismInfo
   * 
   * @author Chris Sachs
   * 
   * @tparam Ring  the element type of the ring.
   * 
-  * @define IntMorphismInfo
-  * A `$Element` has a morphism from `Int` values to $element values.
-  * An `Int` value ''n'' represents the ''nth'' addition of the `$Element`'s
+  * @define LongMorphismInfo
+  * A `$Element` has a morphism from `Long` values to $element values.
+  * An `Long` value ''n'' represents the ''nth'' addition of the `$Element`'s
   * multiplicative identity to its additive identity.
   * 
   * @define vector    $element
@@ -27,19 +27,19 @@ package basis.algebra
   * @define element   `Ring` value
   */
 trait Ring[Ring] extends Vector[Ring, Ring] {
-  /** Adds an `Int` value to this $element.
+  /** Adds an `Long` value to this $element.
     * 
-    * @param  n   the `Int` value to add.
-    * @return the sum of this $element and the `Int` value.
+    * @param  n   the `Long` value to add.
+    * @return the sum of this $element and the `Long` value.
     */
-  def + (n: Int): Ring
+  def + (n: Long): Ring
   
-  /** Subtracts an `Int` value from this $element.
+  /** Subtracts an `Long` value from this $element.
     * 
-    * @param  n   the `Int` value to subtract.
-    * @return the difference of this $element and the `Int` value.
+    * @param  n   the `Long` value to subtract.
+    * @return the difference of this $element and the `Long` value.
     */
-  def - (n: Int): Ring
+  def - (n: Long): Ring
   
   /** Multiplies this $element by another $element.
     * 
@@ -48,19 +48,19 @@ trait Ring[Ring] extends Vector[Ring, Ring] {
     */
   def * (that: Ring): Ring
   
-  /** Multiplies this $element by an `Int` value.
+  /** Multiplies this $element by an `Long` value.
     * 
-    * @param  n   the `Int` value to multiply by.
-    * @return the product of this $element and the `Int` value.
+    * @param  n   the `Long` value to multiply by.
+    * @return the product of this $element and the `Long` value.
     */
-  def * (n: Int): Ring
+  def * (n: Long): Ring
   
-  /** Raises this $element to a positive `Int` power.
+  /** Raises this $element to a positive `Long` power.
     * 
-    * @param  n   the `Int` exponent.
+    * @param  n   the `Long` exponent.
     * @return this $element multiplied by itself `n` times.
     */
-  def pow(n: Int): Ring
+  def pow(n: Long): Ring
   
   def :* (that: Ring): Ring = this * that
